完善登录接口，数据统计接口。

+++ tliasWebAI/src/main/java/com/tlias/service/impl/DeptServiceImpl.java
@@ -22,8 +22,14 @@ public class DeptServiceImpl implements DeptService {
     }
 
     @Override
-    public void deleteById(Integer id) {
-        deptMapper.deleteById(id);
+    public Integer deleteById(Integer id) {
+        if(0 == (Integer) deptMapper.countEmployeesByDeptId(id)){
+            deptMapper.deleteById(id);
+        }
+        else {
+            return 1;
+        }
+        return null;
     }
 
     @Override
